refactor: merge split kernel into single file + address review feedback
- Merge ragged_paged_attention_v3_split_packed.py into
  ragged_paged_attention_v3_split.py (1332 lines, consistent with
  fused kernel single-file pattern)
- Delete ragged_paged_attention_v3_split_packed.py
- Revert utils.py to remove split-only helpers (local test scripts
  are untracked)

Review fixes:
- can_use_split_kv_cache: add per_device_heads % 2 == 0 check
- profile_max_num_token: use getattr for v_head_dim fallback
- _set_kv_cache_after_forward: add tuple length assertion
